+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Fix: Some Fluff-only keywords were being dropped from macro definitions
 -   Fix: Super accessor was not allowing all valid identifiers as a prefix
+-   Fix: Preprocessor did not recognize directives with a space after the `#` symbol.
 
 ## 1.0.11